Refactor plugin into controller classes
