+++ PetIA-app-bridge/composer.json
@@ -1,13 +1,14 @@
 {
-    "name": "petia/app-bridge",
-    "description": "Dependencies for the PetIA App Bridge WordPress plugin",
-    "type": "wordpress-plugin",
-    "autoload": {
-        "psr-4": {
-            "PetIA\\": "includes/"
-        }
-    },
-    "require": {
-        "firebase/php-jwt": "^6.10"
+  "name": "petia/app-bridge",
+  "description": "Dependencies for the PetIA App Bridge WordPress plugin",
+  "type": "wordpress-plugin",
+  "autoload": {
+    "psr-4": {
+      "PetIA\\": "includes/",
+      "PetIA\\Controllers\\": "includes/Controllers/"
     }
+  },
+  "require": {
+    "firebase/php-jwt": "^6.10"
+  }
 }
